Add a fallback NaiveSearchService when ICU components are not compatible with the device
